feat: user upload image endpoint

+++ utils/imageHelper.js
@@ -1,13 +1,15 @@
 const { storage } = require('../config/firebase');
+const DEFAULT_IMAGE_URL = "https://firebasestorage.googleapis.com/v0/b/dropbox-clone-736fa.appspot.com/o/users%2Fuser_2aNC9F2HWDn5x5KjMBi9Y9ywEQX%2Ffiles%2Fdefaultprofile.png?alt=media"
 
 async function uploadImage(fileBuffer, originalFileName, userId, oldImageUrl) {
   const bucket = storage.bucket();
 
-  if (oldImageUrl) {
+  
+  if (oldImageUrl && oldImageUrl !== DEFAULT_IMAGE_URL) {
     await deleteImage(oldImageUrl);
   }
 
-  // Unique name for each image
+// Unique name for each image
   const timestamp = new Date().toISOString();
   const uniqueFileName = `${userId}_${timestamp}_${originalFileName}`;
 
@@ -30,6 +32,11 @@ async function uploadImage(fileBuffer, originalFileName, userId, oldImageUrl) {
 }
 
 async function deleteImage(imageUrl) {
+
+  if (imageUrl === DEFAULT_IMAGE_URL) {
+    return;
+  }
+
   const bucket = storage.bucket();
   const fileName = imageUrl.split('/').pop();
   const file = bucket.file(fileName);
